pyproject.toml
[project]
name = "mcpdoc"
version = "0.0.8"
description = "Server llms-txt documentation over MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "markdownify>=1.1.0",
    "mcp[cli]>=1.4.1",
    "pyyaml>=6.0.1",
    "exa_py>=0.4.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
mcpdoc = "mcpdoc.cli:main"
crawl-exa-docs = "mcpdoc.crawl_exa_docs:main"

[dependency-groups]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.9.7",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "8.0"
# -ra: Report all extra test outcomes (passed, skipped, failed, etc.)
# -q: Enable quiet mode for less cluttered output
# -v: Enable verbose output to display detailed test names and statuses
# --durations=5: Show the 10 slowest tests after the run (useful for performance tuning)
addopts = "-ra -q -v --durations=5"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

